+++ CHANGELOG.md
@@ -6,4 +6,8 @@
 
 - `sandbox#touchp()` and `sandbox#mkdirp()` now also accept an array of paths and create and return all specified files/directories.
 
-- files now have content 
+- files now have content
+
+### Non breaking changes
+
+- `sandbox#touchp()` now accepts a second argument for the content of the file(s)